redirects to the new event
